CED-1730: Streaming + async uploading (#649)

+++ internal/cedana/streamer/dump_adapters.go
@@ -3,17 +3,21 @@ package streamer
 import (
 	"context"
 	"errors"
+	"fmt"
+	"io"
 	"os"
 	"path/filepath"
+	"sync"
 
 	"buf.build/gen/go/cedana/cedana/protocolbuffers/go/daemon"
 	criu_proto "buf.build/gen/go/cedana/criu/protocolbuffers/go/criu"
 	"github.com/cedana/cedana/internal/cedana/filesystem"
 	"github.com/cedana/cedana/pkg/config"
-	"github.com/cedana/cedana/pkg/io"
+	cedana_io "github.com/cedana/cedana/pkg/io"
 	"github.com/cedana/cedana/pkg/plugins"
 	"github.com/cedana/cedana/pkg/profiling"
 	"github.com/cedana/cedana/pkg/types"
+	"github.com/rs/zerolog/log"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
@@ -31,10 +35,13 @@ func DumpFilesystem(streams int32) types.Adapter[types.Dump] {
 			}
 
 			// Check if compression is valid, because we don't want to fail after the dump
-			// as the process would be killed
-			if _, ok := io.SUPPORTED_COMPRESSIONS[compression]; !ok {
+			// as the process would be killed.
+
+			if _, ok := cedana_io.SUPPORTED_COMPRESSIONS[compression]; !ok {
 				return nil, status.Errorf(codes.Unimplemented, "unsupported compression format '%s'", compression)
 			}
+
+			async := config.Global.Checkpoint.Async && storage.IsRemote()
 
 			// If remote storage, we instead use a temporary directory for CRIU
 			if storage.IsRemote() {
@@ -54,11 +61,11 @@ func DumpFilesystem(streams int32) types.Adapter[types.Dump] {
 				return nil, status.Errorf(codes.Internal, "failed to create dump dir: %v", err)
 			}
 			defer func() {
-				if err != nil || storage.IsRemote() {
+				if err != nil || (storage.IsRemote() && !async) {
 					os.RemoveAll(imagesDirectory)
 				}
 			}()
-			err = os.Chmod(imagesDirectory, filesystem.DUMP_DIR_PERMS) // XXX: Because for some reason mkdir is not applying perms
+			err = os.Chmod(imagesDirectory, filesystem.DUMP_DIR_PERMS)
 			if err != nil {
 				return nil, status.Errorf(codes.Internal, "failed to chmod dump dir: %v", err)
 			}
@@ -99,13 +106,24 @@ func DumpFilesystem(streams int32) types.Adapter[types.Dump] {
 
 			path := req.Dir + "/" + req.Name // do not use filepath.Join as it removes a slash
 
+			var streamStorage cedana_io.Storage
+			var storagePath string
+
+			if async {
+				streamStorage = &filesystem.Storage{}
+				storagePath = imagesDirectory
+			} else {
+				streamStorage = storage
+				storagePath = path
+			}
+
 			var waitForIO func() error
 			opts.DumpFs, waitForIO, err = NewStreamingFs(
 				ctx,
 				imgStreamer.BinaryPaths()[0],
 				imagesDirectory,
-				storage,
-				path,
+				streamStorage,
+				storagePath,
 				streams,
 				WRITE_ONLY,
 				compression,
@@ -116,13 +134,84 @@ func DumpFilesystem(streams int32) types.Adapter[types.Dump] {
 
 			// XXX: We do not differentiate between leave-running or not, because unfortunately CRIU
 			// does not close the streaming file descriptors on its side when the PostDumpFunc is triggered.
-			// This is why the logic here is not the same as that in `filesystem/dump_adapters.go`.
+			// This is why the logic here is not the same as that in `filesystem/dump_adapters.go`
+			if async {
+				ext, _ := cedana_io.ExtForCompression(compression)
 
-			defer func() {
-				_, end := profiling.StartTimingCategory(ctx, "storage", waitForIO)
-				err = errors.Join(err, waitForIO())
-				end()
-			}()
+				upload := func(ctx context.Context) error {
+					var wg sync.WaitGroup
+					errCh := make(chan error, streams)
+
+					for i := range streams {
+						wg.Add(1)
+						go func(i int32) {
+							defer wg.Done()
+
+							localPath := fmt.Sprintf("%s/img-%d%s", imagesDirectory, i, ext)
+							remotePath := fmt.Sprintf("%s/img-%d%s", path, i, ext)
+
+							src, err := streamStorage.Open(ctx, localPath)
+							if err != nil {
+								errCh <- fmt.Errorf("failed to open local shard %d: %w", i, err)
+								return
+							}
+							defer src.Close()
+
+							dst, err := storage.Create(ctx, remotePath)
+							if err != nil {
+								errCh <- fmt.Errorf("failed to create remote shard %d: %w", i, err)
+								return
+							}
+							defer dst.Close()
+
+							// add profiling for io copier
+							dst = profiling.IOCategory(ctx, dst, "storage", io.Copy)
+
+							if _, err := io.Copy(dst, src); err != nil {
+								errCh <- fmt.Errorf("failed to upload shard %d: %w", i, err)
+							}
+						}(i)
+					}
+
+					wg.Wait()
+					close(errCh)
+
+					var uploadErr error
+					for e := range errCh {
+						uploadErr = errors.Join(uploadErr, e)
+					}
+
+					os.RemoveAll(imagesDirectory)
+					return uploadErr
+				}
+
+				defer func() {
+					err = errors.Join(err, waitForIO())
+					if err != nil {
+						return
+					}
+
+					// Use a detached context for async upload since the parent request
+					// context will be canceled after the dump completes.
+					uploadCtx := context.WithoutCancel(ctx)
+
+					opts.WG.Add(1)
+					go func() {
+						defer opts.WG.Done()
+						log.Info().Msg("async dump upload started")
+						if uploadErr := upload(uploadCtx); uploadErr != nil {
+							log.Error().Err(uploadErr).Msg("async upload failed")
+						}
+					}()
+				}()
+			} else {
+				// Sync upload, wait for IO completion in PostDumpFunc
+				defer func() {
+					_, end := profiling.StartTimingCategory(ctx, "storage", waitForIO)
+					err = errors.Join(err, waitForIO())
+					end()
+				}()
+			}
 
 			resp.Paths = append(resp.Paths, path)
